Pulls  and displays gifs of doppler radar
Ensure you have mpv installed. This script will download a gif from various websites depending upon your country selection and will display it to the user. I have created a .desktop file for my system and got a random weather icon + aliased it in my .bashrc folder when i run 'doppler' it runs the script or when i click on the .desktop application.